+++ pkg/storage/metupd/update/non/timeline/verify_test.go
@@ -31,7 +31,7 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -52,7 +52,7 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -73,7 +73,7 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -95,7 +95,7 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -136,13 +136,13 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "x",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -165,13 +165,13 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -194,21 +194,21 @@ func Test_Timeline_Verify_Input_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "x",
-								Value: []int64{
+								Value: []float64{
 									73,
 									91,
 								},
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									22,
 									94,
 								},
 							},
 							{
 								Space: "t",
-								Value: []int64{
+								Value: []float64{
 									20,
 									16,
 								},
@@ -266,7 +266,7 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -302,7 +302,7 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -324,7 +324,7 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -347,7 +347,7 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 									1,
@@ -373,14 +373,14 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "x",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -403,21 +403,21 @@ func Test_Timeline_Verify_Input_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "x",
-								Value: []int64{
+								Value: []float64{
 									73,
 									91,
 								},
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									22,
 									94,
 								},
 							},
 							{
 								Space: "z",
-								Value: []int64{
+								Value: []float64{
 									20,
 									16,
 								},
@@ -476,7 +476,7 @@ func Test_Timeline_Verify_Search_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -503,7 +503,7 @@ func Test_Timeline_Verify_Search_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 									1,
@@ -532,7 +532,7 @@ func Test_Timeline_Verify_Search_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -558,7 +558,7 @@ func Test_Timeline_Verify_Search_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -585,7 +585,7 @@ func Test_Timeline_Verify_Search_False(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "toolong",
-								Value: []int64{
+								Value: []float64{
 									32,
 								},
 							},
@@ -650,7 +650,7 @@ func Test_Timeline_Verify_Search_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 								},
@@ -677,7 +677,7 @@ func Test_Timeline_Verify_Search_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 									1,
@@ -706,7 +706,7 @@ func Test_Timeline_Verify_Search_True(t *testing.T) {
 						Data: []*metupd.UpdateI_Obj_Property_Data{
 							{
 								Space: "x",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 									1,
@@ -715,7 +715,7 @@ func Test_Timeline_Verify_Search_True(t *testing.T) {
 							},
 							{
 								Space: "y",
-								Value: []int64{
+								Value: []float64{
 									32,
 									85,
 									1,
