+++ pkg/value/metric/timeline/data/spec.go
@@ -2,5 +2,5 @@ package data
 
 type Interface interface {
 	GetSpace() string
-	GetValue() []int64
+	GetValue() []float64
 }
